pyproject.toml
[tool.poetry]
name = "dms-downloader"
version = "0.1.0"
description = ""
authors = ["Hirusha Fernando <heshanhfernando@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
redis = "^5.2.1"
tqdm = "^4.67.1"
xmltodict = "^0.14.2"
python-dotenv = "^1.1.0"
fastapi = "^0.115.12"
bs4 = "^0.0.2"
uvicorn = "^0.34.0"
textual = "^3.0.1"
python-multipart = "^0.0.20"
jinja2 = "^3.1.6"
websockets = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
